- Added using of the .env file - New SQL scripts to support filtering - Python package 0.0.2 release - Improved links sent inside Slack messages - README updates - Refactored not to use init_env_vars.sh inside scripts (issue with passing through selected environment otherwise), but to check if env was properly set as a standalone action.

+++ utils/bq_cleanup.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-source "$DIR/../setup/init_env_vars.sh"
+bash "$DIR/../setup/check_setup.sh"
+retVal=$?
+if [ $retVal -eq 2 ]; then
+  exit 2
+fi
 
 DELETE='DELETE FROM `'"${PROJECT_ID}"'`.'"${DATASET}."''"${TASK_STATUS_TABLE_ID}"' WHERE true; '
